Create `CpuCalculator`
While leaving a kind of mess lol
