--- Mediapipe.Net/Calculators/IGpuCalculator.cs
@@ -1,11 +1,0 @@
-// Copyright (c) homuler and Vignette
-// This file is part of MediaPipe.NET.
-// MediaPipe.NET is licensed under the MIT License. See LICENSE for details.
-
-namespace Mediapipe.Net.Calculators
-{
-    public interface IGpuCalculator<T> : ICalculator<T>
-    {
-
-    }
-}
